Former-commit-id: 0bb71e1d7bdd0442be91519c037b1890dd0f4bda

+++ Tools/QuickEd/Classes/Model/ControlProperties/StyleSheetProperty.cpp
@@ -69,9 +69,9 @@ StyleSheetProperty::~StyleSheetProperty()
     properties.clear();
 }
 
-int StyleSheetProperty::GetCount() const
+uint32 StyleSheetProperty::GetCount() const
 {
-    return static_cast<int>(properties.size());
+    return properties.size();
 }
 
 AbstractProperty *StyleSheetProperty::GetProperty(int index) const
